% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddDataSets.R
\docType{package}
\name{ddDataSets}
\alias{ddDataSets}
\alias{ddDataSets-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
ddDataSets
A data package for ddDataSets.

Use \code{data(package='ddDataSets')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{tidy_iris}
}
