% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddDataSets.R
\docType{data}
\name{tidy_cars}
\alias{tidy_cars}
\title{Descriptive data title}
\format{a \code{spec_tbl_df} containing the following fields:
\describe{
\item{speed_mph}{}
\item{stopping_distace}{}
}}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{ddDataSets}
}
