% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddDataSets.R
\docType{data}
\name{tidy_iris}
\alias{tidy_iris}
\title{Descriptive data title}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{Sepal.Length}{}
\item{Sepal.Width}{}
\item{Petal.Length}{}
\item{Petal.Width}{}
\item{Species}{}
}}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{ddDataSets}
}
